+++ Back-end/muut/src/main/java/com/fp/muut/login/CustomerRepository.java
@@ -5,30 +5,45 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import com.fp.muut.entity.Customer;
 
+import jakarta.persistence.EntityManager;
+
 @Repository
 public class CustomerRepository {
 
-	private static Map<Long, Customer> store = new HashMap<>();
-	private static long sequence = 0L;
-	
-	public Customer save(Customer customer) {
-		customer.setCustomerNum(++sequence);
-		store.put(customer.getCustomerNum(), customer);
-		return customer;
+	@Autowired
+	private EntityManager em;
+
+	//저장
+	public void join(Customer customer) {
+		em.persist(customer);
 	}
 	
-	public Customer findById(Long id) {
-		return store.get(id);
+	//조회
+	public Customer findOne(String customerId) {
+		return em.find(Customer.class, customerId);
 	}
-	
+
+	//전체조회
 	public List<Customer> findAll(){
-		return new ArrayList<>(store.values());
+		return em.createQuery("select c from Customer c", Customer.class).getResultList();
 	}
 	
+//	//이름으로 조회
+//	public List<Customer> findByName(String customer_name){
+//		return em.createQuery("select c from Customer c where c.customer_name = :customer_name", Customer.class).setParameter("customerNmae", customer_name).getResultList();
+//	}
+	
+	//아이디로 조회
+		public List<Customer> findById(String customer_id){
+			return em.createQuery("select c from Customer c where c.customer_id = :customer_id", Customer.class).setParameter("customerId", customer_id).getResultList();
+		}
+
+	//로그인
 	public Customer findByLoginId(String loginId) {
 		List<Customer> all = findAll();
 		for (Customer m : all) {
